full react: using widgets instead helper methods
