+++ lib/imgy_utils.dart
@@ -1,23 +1,21 @@
-part of './imgy.dart';
+import 'dart:math';
 
-extension ImgyUtils on ImgyState {
-  checkSrc() {
-    return Uri.parse(widget.src).isAbsolute;
+checkSrc(String src) {
+  return Uri.parse(src).isAbsolute;
+}
+
+checkFullSrc(String src) {
+  return Uri.parse(src).isAbsolute;
+}
+
+String randomString(int length) {
+  const chars = "abcdefghijklmnopqrstuvwxyz0123456789";
+  final rnd = Random(DateTime.now().millisecondsSinceEpoch);
+  final buf = StringBuffer();
+
+  for (var x = 0; x < length; x++) {
+    buf.write(chars[rnd.nextInt(chars.length)]);
   }
 
-  checkFullSrc() {
-    return Uri.parse(widget.fullSrc).isAbsolute;
-  }
-
-  String randomString(int length) {
-    const chars = "abcdefghijklmnopqrstuvwxyz0123456789";
-    final rnd = Random(DateTime.now().millisecondsSinceEpoch);
-    final buf = StringBuffer();
-
-    for (var x = 0; x < length; x++) {
-      buf.write(chars[rnd.nextInt(chars.length)]);
-    }
-
-    return buf.toString();
-  }
+  return buf.toString();
 }
